mutslice force nillability
